refactor: UI 코드에서 ColorDTO 뜯어내고 LectureColor 사용하기 (#482)

+++ app/src/main/java/com/wafflestudio/snutt2/lib/network/dto/core/LectureDto.kt
@@ -184,7 +184,7 @@ data class LectureDto(
                     BuiltInColor(
                         foreground = Color(color.fgRaw?.toColorInt() ?: 0xFFFFFF),
                         background = Color(color.bgRaw?.toColorInt() ?: 0xFFFFFF),
-                        colorIndex = colorIndex.toInt(),
+                        colorIndex = colorIndex,
                     )
                 },
             )
@@ -214,7 +214,7 @@ data class LectureDto(
                     BuiltInColor(
                         foreground = Color(color.fgRaw?.toColorInt() ?: 0xFFFFFF),
                         background = Color(color.bgRaw?.toColorInt() ?: 0xFFFFFF),
-                        colorIndex = colorIndex.toInt(),
+                        colorIndex = colorIndex,
                     )
                 },
             )
